Move feature flag files
