+++ web_ui/packages/core/src/feature-flags/hooks/use-feature-flag-query.hook.ts
@@ -1,12 +1,12 @@
 // Copyright (C) 2022-2025 Intel Corporation
 // LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
 
-import { useApplicationServices } from '@geti/core/src/services/application-services-provider.component';
 import { useSuspenseQuery, UseSuspenseQueryOptions, UseSuspenseQueryResult } from '@tanstack/react-query';
 import { AxiosError } from 'axios';
 import { isEmpty } from 'lodash-es';
 
-import QUERY_KEYS from '../../../../packages/core/src/requests/query-keys';
+import QUERY_KEYS from '../../requests/query-keys';
+import { useApplicationServices } from '../../services/application-services-provider.component';
 import { FeatureFlags, FeatureFlagService } from '../services/feature-flag-service.interface';
 
 const featureFlagQueryOptions = (
