+++ web_ui/packages/core/src/feature-flags/services/api-feature-flag-service.test.ts
@@ -3,9 +3,9 @@
 
 import { rest } from 'msw';
 
-import { apiRequestUrl } from '../../../../packages/core/src/services/test-utils';
-import { API_URLS } from '../../../../packages/core/src/services/urls';
-import { server } from '../../annotations/services/test-utils';
+import { server } from '../../../../../src/core/annotations/services/test-utils';
+import { apiRequestUrl } from '../../services/test-utils';
+import { API_URLS } from '../../services/urls';
 import { createApiFeatureFlagService } from './api-feature-flag-service';
 
 describe('FeatureFlagService', () => {
